requirements.txt
PyQt6~=6.7.0
transformers~=4.41.2
torch~=2.3.1
pillow~=10.3.0